Rmp 4 ticket 01 place owner and rating precision

+++ src/graphql/typeDefs/placeTypedefs.ts
@@ -22,7 +22,7 @@ export const placeDefs: DocumentNode = gql`
         phone: String
         website: String
         category: String
-        averageRating: Int
+        averageRating: Float
         reviewCount: Int
         isVerified: Boolean
     }
